@@ -16,7 +16,7 @@ import 'dart:ui';
  Developer Notes
 
  SUPPORT EMAIL - support@woosignal.com
- VERSION - 2.6.0
+ VERSION - 3.0.0
  https://woosignal.com
  */
 
@@ -29,19 +29,7 @@ const app_key = "Your app key from WooSignal";
 // Your App key from WooSignal
 // link: https://woosignal.com/dashboard/apps
 
-const app_logo_url = "https://woosignal.com/images/120x120_woosignal.png";
-
-const app_terms_url = "https://yourdomain.com/terms";
-const app_privacy_url = "https://yourdomain.com/privacy";
-
 /*<! ------ APP SETTINGS ------!>*/
-
-const app_currency_symbol = "\£";
-const app_currency_iso = "gbp";
-
-const app_products_prices_include_tax = true;
-
-const app_disable_shipping = false;
 
 const Locale app_locale = Locale('en');
 
@@ -57,7 +45,8 @@ const List<Locale> app_locales_supported = [
 // then create a new lang json file using keys from en.json
 // e.g. lang/es.json
 
-const app_product_placeholder_image = "https://woosignal.com/images/woocommerce-placeholder.png";
+const app_product_placeholder_image =
+    "https://woosignal.com/images/woocommerce-placeholder.png";
 
 /*<! ------ PAYMENT GATEWAYS ------!>*/
 
@@ -73,24 +62,11 @@ const app_payment_methods = ["Stripe"];
 
 const app_stripe_account = "Your Stripe Key from WooSignal";
 
-const app_stripe_live_mode = false;
+const app_stripe_live_mode = false; // set to true for live Stripe payments
 // For Live Payments follow the below steps
 // #1 SET the above to true for live payments
 // #2 Next visit https://woosignal.com/dashboard
 // #3 Then change "Environment for Stripe" to Live mode
-
-/*<! ------ WP LOGIN (OPTIONAL) ------!>*/
-
-// Allows customers to login/register, view account, purchase items as a user.
-// #1 Install the "WP JSON API" plugin on WordPress via https://woosignal.com/plugins/wordpress/wp-json-api
-// #2 Next activate the plugin on your WordPress and enable "use_wp_login = true"
-// link: https://woosignal.com/dashboard/plugins
-
-const use_wp_login = false;
-const app_base_url = "https://mysite.com"; // change to your url
-const app_forgot_password_url =
-    "https://mysite.com/my-account/lost-password"; // change to your forgot password url
-const app_wp_api_path = "/wp-json"; // By default "/wp-json" should work
 
 /*<! ------ Razor Pay (OPTIONAL) ------!>*/
 // https://razorpay.com/
