@@ -32,31 +32,19 @@ class _AccountBillingDetailsPageState extends State<AccountBillingDetailsPage> {
   _AccountBillingDetailsPageState();
 
   // BILLING TEXT CONTROLLERS
-  TextEditingController _txtShippingFirstName;
-  TextEditingController _txtShippingLastName;
-  TextEditingController _txtShippingAddressLine;
-  TextEditingController _txtShippingCity;
-  TextEditingController _txtShippingState;
-  TextEditingController _txtShippingPostalCode;
-  TextEditingController _txtShippingCountry;
+  TextEditingController _txtShippingFirstName = TextEditingController();
+  TextEditingController _txtShippingLastName = TextEditingController();
+  TextEditingController _txtShippingAddressLine = TextEditingController();
+  TextEditingController _txtShippingCity = TextEditingController();
+  TextEditingController _txtShippingState = TextEditingController();
+  TextEditingController _txtShippingPostalCode = TextEditingController();
+  TextEditingController _txtShippingCountry = TextEditingController();
 
-  bool _isLoading, _isUpdating;
+  bool _isLoading = true, _isUpdating = false;
 
   @override
   void initState() {
     super.initState();
-
-    _txtShippingFirstName = TextEditingController();
-    _txtShippingLastName = TextEditingController();
-    _txtShippingAddressLine = TextEditingController();
-    _txtShippingCity = TextEditingController();
-    _txtShippingState = TextEditingController();
-    _txtShippingPostalCode = TextEditingController();
-    _txtShippingCountry = TextEditingController();
-
-    _isLoading = true;
-    _isUpdating = false;
-
     _fetchUserDetails();
   }
 
@@ -84,7 +72,7 @@ class _AccountBillingDetailsPageState extends State<AccountBillingDetailsPage> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      resizeToAvoidBottomPadding: false,
+      resizeToAvoidBottomInset: false,
       appBar: AppBar(
         backgroundColor: Colors.transparent,
         title: Text(
@@ -113,16 +101,14 @@ class _AccountBillingDetailsPageState extends State<AccountBillingDetailsPage> {
                               Row(
                                 children: <Widget>[
                                   Flexible(
-                                    child: wsTextEditingRow(
-                                      context,
+                                    child: TextEditingRow(
                                       heading: trans(context, "First Name"),
                                       controller: _txtShippingFirstName,
                                       shouldAutoFocus: true,
                                     ),
                                   ),
                                   Flexible(
-                                    child: wsTextEditingRow(
-                                      context,
+                                    child: TextEditingRow(
                                       heading: trans(context, "Last Name"),
                                       controller: _txtShippingLastName,
                                     ),
@@ -132,21 +118,19 @@ class _AccountBillingDetailsPageState extends State<AccountBillingDetailsPage> {
                                 mainAxisAlignment:
                                     MainAxisAlignment.spaceEvenly,
                               ),
-                              wsTextEditingRow(
-                                context,
+                              TextEditingRow(
                                 heading: trans(context, "Address Line"),
                                 controller: _txtShippingAddressLine,
                               ),
                               Row(children: <Widget>[
                                 Flexible(
-                                  child: wsTextEditingRow(
-                                    context,
+                                  child: TextEditingRow(
                                     heading: trans(context, "City"),
                                     controller: _txtShippingCity,
                                   ),
                                 ),
                                 Flexible(
-                                  child: wsTextEditingRow(context,
+                                  child: TextEditingRow(
                                       heading: trans(context, "State"),
                                       keyboardType: TextInputType.emailAddress,
                                       controller: _txtShippingState),
@@ -155,14 +139,13 @@ class _AccountBillingDetailsPageState extends State<AccountBillingDetailsPage> {
                               Row(
                                 children: <Widget>[
                                   Flexible(
-                                    child: wsTextEditingRow(
-                                      context,
+                                    child: TextEditingRow(
                                       heading: trans(context, "Postal code"),
                                       controller: _txtShippingPostalCode,
                                     ),
                                   ),
                                   Flexible(
-                                    child: wsTextEditingRow(context,
+                                    child: TextEditingRow(
                                         heading: trans(context, "Country"),
                                         keyboardType:
                                             TextInputType.emailAddress,
@@ -198,10 +181,10 @@ class _AccountBillingDetailsPageState extends State<AccountBillingDetailsPage> {
                       ),
                       Column(
                         children: <Widget>[
-                          wsPrimaryButton(context,
+                          PrimaryButton(
                               title: trans(context, "UPDATE DETAILS"),
                               action:
-                                  _isUpdating ? null : _updateBillingDetails),
+                                  _isUpdating ? () {} : _updateBillingDetails),
                         ],
                       ),
                     ],
